refactor: allow parts of the code to be used as library
