+++ pkg/signature/signature_opts.go
@@ -1,4 +1,4 @@
-package vault
+package signature
 
 import "errors"
